Modified according to code review.
