+++ CefSharp.Wpf/CefSharp.Wpf.csproj
@@ -79,16 +79,17 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CefSettings.cs" />
-    <Compile Include="IME\IMEHandler.cs" />
-    <Compile Include="IME\NativeIME.cs" />    
-    <Compile Include="Experimental\ChromiumWebBrowserWithTouchSupport.cs" />    
+    <Compile Include="Experimental\ChromiumWebBrowserWithTouchSupport.cs" />
     <Compile Include="Internals\DragOperationMaskExtensions.cs" />
+    <Compile Include="Internals\IMEHandler.cs" />
     <Compile Include="Internals\MonitorInfo.cs" />
     <Compile Include="Internals\MonitorInfoEx.cs" />
+    <Compile Include="Internals\NativeIME.cs" />
     <Compile Include="Internals\RectStruct.cs" />
+    <Compile Include="Internals\WpfIMEKeyboardHandler.cs" />
+    <Compile Include="Internals\WpfKeyboardHandler.cs" />
     <Compile Include="IWpfKeyboardHandler.cs" />
     <Compile Include="Internals\VirtualKeys.cs" />
-    <Compile Include="Internals\WpfKeyboardHandler.cs" />
     <Compile Include="Internals\WpfLegacyKeyboardHandler.cs" />
     <Compile Include="DelegateCommand.cs" />
     <Compile Include="IRenderHandler.cs" />
