+++ CefSharp.Wpf/Internals/IMEHandler.cs
@@ -1,12 +1,18 @@
+// Copyright © 2018 The CefSharp Authors. All rights reserved.
+//
+// Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
+
 using System;
 using System.Collections.Generic;
 using System.Text;
 using CefSharp.Structs;
-using static CefSharp.Wpf.IME.NativeIME;
 
-namespace CefSharp.Wpf.IME
+namespace CefSharp.Wpf.Internals
 {
-    public class IMEHandler : IDisposable
+    /// <summary>
+    /// IMEHandler provides implementattion when message WM_IME_COMPOSITION is received.
+    /// </summary>
+    public class IMEHandler
     {
         // Black SkColor value for underline.
         internal const uint ColorUNDERLINE = 0xFF000000;
@@ -15,74 +21,69 @@ namespace CefSharp.Wpf.IME
 
         IntPtr hWnd;
 
-        internal static IMEHandler Create(IntPtr hWnd)
-        {
-            return new IMEHandler(hWnd);
-        }
-
-        private IMEHandler(IntPtr hWnd)
+        public IMEHandler(IntPtr hWnd)
         {
             this.hWnd = hWnd;
         }
 
-        public void Dispose()
-        {
-        }
-
-        private bool GetString(IntPtr hIMC, uint lParam, uint type, out string text)
-        {
-            text = string.Empty;
-
-            if (!IsParam(lParam, type))
-                return false;
-
-            var strLen = ImmGetCompositionString(hIMC, type, null, 0);
-            if (strLen <= 0)
-                return false;
-
-            // buffer contains char (2 bytes)
-            byte[] buffer = new byte[strLen];
-            ImmGetCompositionString(hIMC, type, buffer, strLen);
-            text = Encoding.Unicode.GetString(buffer);
-
-            return true;
-        }
-
         internal bool GetResult(uint lParam, out string text)
         {
-            IntPtr hIMC = ImmGetContext(hWnd);
+            IntPtr hIMC = NativeIME.ImmGetContext(hWnd);
 
-            var ret = GetString(hIMC, lParam, GCS_RESULTSTR, out text);
+            var ret = GetString(hIMC, lParam, NativeIME.GCS_RESULTSTR, out text);
 
-            ImmReleaseContext(hWnd, hIMC);
+            NativeIME.ImmReleaseContext(hWnd, hIMC);
 
             return ret;
         }
 
         internal bool GetComposition(uint lParam, List<CompositionUnderline> underlines, ref int compositionStart, out string text)
         {
-            IntPtr hIMC = ImmGetContext(hWnd);
+            IntPtr hIMC = NativeIME.ImmGetContext(hWnd);
 
-            bool ret = GetString(hIMC, lParam, GCS_COMPSTR, out text);
+            bool ret = GetString(hIMC, lParam, NativeIME.GCS_COMPSTR, out text);
             if (ret)
             {
                 GetCompositionInfo(lParam, text, underlines, ref compositionStart);
             }
 
-            ImmReleaseContext(hWnd, hIMC);
+            NativeIME.ImmReleaseContext(hWnd, hIMC);
 
             return ret;
         }
 
+        private static bool GetString(IntPtr hIMC, uint lParam, uint type, out string text)
+        {
+            text = string.Empty;
+
+            if (!IsParam(lParam, type))
+            {
+                return false;
+            }
+
+            var strLen = NativeIME.ImmGetCompositionString(hIMC, type, null, 0);
+            if (strLen <= 0)
+            {
+                return false;
+            }
+
+            // buffer contains char (2 bytes)
+            byte[] buffer = new byte[strLen];
+            NativeIME.ImmGetCompositionString(hIMC, type, buffer, strLen);
+            text = Encoding.Unicode.GetString(buffer);
+
+            return true;
+        }
+
         private void GetCompositionInfo(uint lParam, string text, List<CompositionUnderline> underlines, ref int compositionStart)
         {
-            IntPtr hIMC = ImmGetContext(hWnd);
+            IntPtr hIMC = NativeIME.ImmGetContext(hWnd);
 
             underlines.Clear();
 
             int targetStart = text.Length;
             int targetEnd = text.Length;
-            if (IsParam(lParam, GCS_COMPATTR))
+            if (IsParam(lParam, NativeIME.GCS_COMPATTR))
             {
                 GetCompositionSelectionRange(hIMC, ref targetStart, ref targetEnd);
             }
@@ -91,11 +92,11 @@ namespace CefSharp.Wpf.IME
             // it means the cursor should not be moved and we therefore place the caret at
             // the beginning of the composition string. Otherwise we should honour the
             // GCS_CURSORPOS value if it's available.
-            if (!IsParam(lParam, CS_NOMOVECARET) && IsParam(lParam, GCS_CURSORPOS))
+            if (!IsParam(lParam, NativeIME.CS_NOMOVECARET) && IsParam(lParam, NativeIME.GCS_CURSORPOS))
             {
                 // IMM32 does not support non-zero-width selection in a composition. So
                 // always use the caret position as selection range.
-                int cursor = (int)ImmGetCompositionString(hIMC, GCS_CURSORPOS, null, 0);
+                int cursor = (int)NativeIME.ImmGetCompositionString(hIMC, NativeIME.GCS_CURSORPOS, null, 0);
                 compositionStart = cursor;
             }
             else
@@ -103,7 +104,7 @@ namespace CefSharp.Wpf.IME
                 compositionStart = 0;
             }
 
-            if (IsParam(lParam, GCS_COMPCLAUSE))
+            if (IsParam(lParam, NativeIME.GCS_COMPCLAUSE))
             {
                 GetCompositionUnderlines(hIMC, targetStart, targetEnd, underlines);
             }
@@ -133,20 +134,22 @@ namespace CefSharp.Wpf.IME
                 underlines.Add(new CompositionUnderline(range, ColorUNDERLINE, ColorBKCOLOR, thick));
             }
 
-            ImmReleaseContext(hWnd, hIMC);
+            NativeIME.ImmReleaseContext(hWnd, hIMC);
         }
 
-        private void GetCompositionUnderlines(IntPtr hIMC, int targetStart, int targetEnd, List<CompositionUnderline> underlines)
+        private static void GetCompositionUnderlines(IntPtr hIMC, int targetStart, int targetEnd, List<CompositionUnderline> underlines)
         {
-            var clauseSize = ImmGetCompositionString(hIMC, GCS_COMPCLAUSE, null, 0);
+            var clauseSize = NativeIME.ImmGetCompositionString(hIMC, NativeIME.GCS_COMPCLAUSE, null, 0);
             if (clauseSize <= 0)
+            {
                 return;
+            }
 
             int clauseLength = (int)clauseSize / sizeof(Int32);
 
             // buffer contains 32 bytes (4 bytes) array
             var clauseData = new byte[(int)clauseSize];
-            ImmGetCompositionString(hIMC, GCS_COMPCLAUSE, clauseData, clauseSize);
+            NativeIME.ImmGetCompositionString(hIMC, NativeIME.GCS_COMPCLAUSE, clauseData, clauseSize);
 
             var clauseLength_1 = clauseLength - 1;
             for (int i = 0; i < clauseLength_1; i++)
@@ -161,41 +164,47 @@ namespace CefSharp.Wpf.IME
             }
         }
 
-        private void GetCompositionSelectionRange(IntPtr hIMC, ref int targetStart, ref int targetEnd)
+        private static void GetCompositionSelectionRange(IntPtr hIMC, ref int targetStart, ref int targetEnd)
         {
-            var attributeSize = ImmGetCompositionString(hIMC, GCS_COMPATTR, null, 0);
-			if (attributeSize <= 0)
+            var attributeSize = NativeIME.ImmGetCompositionString(hIMC, NativeIME.GCS_COMPATTR, null, 0);
+            if (attributeSize <= 0)
+            {
                 return;
+            }
 
             int start = 0;
             int end = 0;
 
             // Buffer contains 8bit array
             var attributeData = new byte[attributeSize];
-            ImmGetCompositionString(hIMC, GCS_COMPATTR, attributeData, attributeSize);
+            NativeIME.ImmGetCompositionString(hIMC, NativeIME.GCS_COMPATTR, attributeData, attributeSize);
 
             for (start = 0; start < attributeSize; ++start)
             {
                 if (IsSelectionAttribute(attributeData[start]))
+                {
                     break;
+                }
             }
 
             for (end = start; end < attributeSize; ++end)
             {
                 if (!IsSelectionAttribute(attributeData[end]))
+                {
                     break;
+                }
             }
 
             targetStart = start;
             targetEnd = end;
         }
 
-        private bool IsSelectionAttribute(byte attribute)
+        private static bool IsSelectionAttribute(byte attribute)
         {
-            return (attribute == ATTR_TARGET_CONVERTED || attribute == ATTR_TARGET_NOTCONVERTED);
+            return (attribute == NativeIME.ATTR_TARGET_CONVERTED || attribute == NativeIME.ATTR_TARGET_NOTCONVERTED);
         }
 
-        internal static bool IsParam(uint lParam, uint type)
+        private static bool IsParam(uint lParam, uint type)
         {
             return (lParam & type) == type;
         }
